+++ reflex-dom-core/src/Reflex/Dom/Class.hs
@@ -4,8 +4,7 @@ module Reflex.Dom.Class ( module Reflex.Dom.Class
                         , module Web.KeyCode
                         ) where
 
-import Data.Map (Map)
-import qualified Data.Map as Map
+import Control.Lens
 import Reflex.Class
 import Web.KeyCode
 
@@ -13,11 +12,9 @@ import Foreign.JavaScript.TH
 import Reflex.PerformEvent.Class
 import Reflex.PostBuild.Class
 
-import Prelude hiding (concat, mapM, mapM_, sequence)
-
--- | Alias for Data.Map.singleton
-(=:) :: k -> a -> Map k a
-(=:) = Map.singleton
+-- | Previously an alias for 'Data.Map.singleton', but now generalised to 'At'
+(=:) :: (At m, Monoid m) => Index m -> IxValue m -> m
+k =: a = at k ?~ a $ mempty
 infixr 7 =: -- Ought to bind tighter than <>, which is infixr 6
 
 {-# DEPRECATED keycodeEnter "Instead of `x == keycodeEnter`, use `keyCodeLookup x == Enter`" #-}