+++ reflex-dom/ChangeLog.md
@@ -1,5 +1,17 @@
 # Revision history for reflex-dom
 
+## 0.6.0.0
+
+* ([#379](https://github.com/reflex-frp/reflex-dom/pull/379)) Re-export all modules from `reflex-dom-core`. The newly re-exported modules are:
+    * `Foreign.JavaScript.Utils`
+    * `Reflex.Dom.Builder.Hydratable`
+    * `Reflex.Dom.Modals.Class`
+    * `Reflex.Dom.Prerender`
+    * `Reflex.Dom.Time`
+    * `Reflex.Dom.WebSocket.Query`
+    * `Reflex.Dom.Xhr.FormData`
+* ([#366](https://github.com/reflex-frp/reflex-dom/pull/366)) Bump bounds for `reflex` to include 0.7.
+
 ## 0.5.3
 
 * On Android, enable prompting the user for geolocation
@@ -18,4 +30,5 @@
 ## 0.5
 
 * Re-export new hydration widget "mainWidget" functions.
+
 * The use-warp flag now properly takes precedence on macOS.